refactor: 板の登録処理をBoardRepositoryに移動
板情報の永続化に関連するロジックを`BookmarkBoardRepository`から`BoardRepository`へ移行しました。
具体的には、`ensureBoard`メソッドを`BoardRepository`に移動し、`BookmarkBoardRepository`では`BoardRepository`の同メソッドを呼び出すように変更しました。
また、`BoardRepository`の`fetchBoardName`メソッドにおいて、板タイトル取得のキーを`BBS_TITLE_ORIG`から`BBS_TITLE`に変更しました。
